[Update] Prints

+++ lib/random_number_genarator.h
@@ -9,9 +9,9 @@
 #include <limits.h>
 
 
-double expo_distribution(double lambda) {
+int expo_distribution(double lambda) {
     double u = rand() / (RAND_MAX + 1.0);
-    return lambda * -log(1.0 - u);
+    return (int)(lambda * -log(1.0 - u));
 }
 
 /* poisson.c
@@ -22,8 +22,8 @@ double expo_distribution(double lambda) {
  *
  * Time complexity is O(lambda), which is not optimal.
 */
-long poisson_distribution(double lambda) {
-    long k = 0;
+int poisson_distribution(double lambda) {
+    int k = 0;
     double L = exp(-lambda), p = 1;
 
     do {
